+++ examples/next/src/app/fixtures/experiments/1-varied-component/README.mdx
@@ -1,6 +1,10 @@
 # Varying a Module
 
+## Explanation
+
 This example shows a component module being varied.
+
+## Activation
 
 If an `experiments` header is set, with a `"test-feature": { "bucket": "test-variant" }`
 decision, the "Control 1" component is replaced by a "Variant 1".
