+++ examples/next/src/app/fixtures/experiments/4-varied-variant/README.mdx
@@ -1,9 +1,13 @@
 # Varying a Variation
 
+## Explanation
+
 This example shows compound variation, in that a component varied by "feature 1"
 can be varied again by a "feature 2".  In this instance, the same point cut
 (for experimentation) is acting on both, since it's `variantGlob` is defined
 with a recursive [`globstar`](https://www.gnu.org/software/bash/manual/html_node/The-Shopt-Builtin.html#:~:text=globstar,only%20directories%20and%20subdirectories%20match.).
+
+## Activation
 
 If an `experiments` header is set, with a `"test-feature": { "bucket": "test-variant" }`
 decision, the "Control 1" component is replaced by a "Variant 1" component.
