+++ examples/next/src/app/fixtures/experiments/4-varied-variant/playwright.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect, ConsoleMessage } from "@playwright/test";
 import setExperimentHeaders from "../playwright.setExperimentHeaders";
 import locateWithinExample from "../playwright.locateInExample";
-import getFixtureURL from "../playwright.getFixtureUrl";
+import getFixtureURL from "../../playwright.getFixtureUrl";
 const fixtureURL = getFixtureURL(import.meta.url);
 
 test.describe("varying a variant in a second experiment", () => {
