+++ examples/next/src/app/fixtures/experiments/5-toggle-config-opt-out/README.mdx
@@ -1,7 +1,11 @@
 # Opting Out of Join Points using a local Toggle Config
 
+## Explanation
+
 This example demonstrates the use of a `__toggleConfig.json` within a folder, 
 restricting all potential join points to the contents of its `joinPoints` array.
+
+## Activation
 
 If an `experiments` header is set, with a `"test-feature": { "bucket": "test-variant" }` decision,
 the control component will still show, since the `__toggleConfig.json` has an empty `joinPoints` array.
