+++ examples/next/src/app/fixtures/experiments/6-toggle-config-variant-filter-same-directory/README.mdx
@@ -1,7 +1,11 @@
 # Filtering Join Points using a Toggle Config
 
+## Explanation
+
 This example demonstrates the use of a `__toggleConfig.json` within a folder,
 filtering the join points to a proscribed list.
+
+## Activation
 
 If an `experiments` header is set, with a `"test-feature": { "bucket": "test-variant" }`
 decision, the "Control 1" (`component.tsx`) component is replaced by a "Variant 1".
