+++ examples/next/src/app/fixtures/experiments/6-toggle-config-variant-filter-same-directory/playwright.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from "@playwright/test";
 import setExperimentHeaders from "../playwright.setExperimentHeaders";
 import locateWithinExample from "../playwright.locateInExample";
-import getFixtureURL from "../playwright.getFixtureUrl";
+import getFixtureURL from "../../playwright.getFixtureUrl";
 const fixtureURL = getFixtureURL(import.meta.url);
 
 test.describe("varying a component whilst using a toggle config to control the join point", () => {
