+++ examples/next/src/app/fixtures/experiments/7-toggle-config-variant-filter-alternate-directory/playwright.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from "@playwright/test";
 import setExperimentHeaders from "../playwright.setExperimentHeaders";
 import locateWithinExample from "../playwright.locateInExample";
-import getFixtureURL from "../playwright.getFixtureUrl";
+import getFixtureURL from "../../playwright.getFixtureUrl";
 const fixtureURL = getFixtureURL(import.meta.url);
 
 test.describe("varying a component and directly referencing a variant dependency within another folder from that variant, using a toggleConfig to prevent that dependency being toggled itself", () => {
